move TrackStatePropMask into separate file
